requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
openai==1.37.0
requests==2.32.3
streamlit==1.36.0
